Move common logic to base class
